requirements.txt
streamlit==1.25.0 
cohere==4.6.0    
PyPDF2==3.0.1
python-docx==0.8.11
